refactor(reporting): move catalog to views/reporting/catalog.py
Split api_field_values, api_metrics, _metric_label,
_accessible_curated_sources, and _labeled_field_values out of
nx_lib/views/reporting/__init__.py into catalog.py, part of
beautify-phase-2a Task 3 -- the last cluster. Also moved
_accessible_sql_targets to run.py (SQL-target helper, not a route,
didn't fit any of the 8 named clusters cleanly; it belongs next to
the other SQL-sandbox helpers). __init__.py re-exports both plus
_labeled_field_values (used directly by tests/unit/test_reporting_scope.py)
and delegates route registration to catalog.register_routes(app).

ai.py's lazy `from . import _accessible_sql_targets` /
`_accessible_curated_sources` (reads the package namespace at call
time, not the submodule) keeps working unchanged since __init__.py
still re-exports both names.

__init__.py is now 173 lines: imports, the re-export __all__ list, and
register_routes's fan-out to all nine submodules (pages, admin_registry,
catalog, run, ai, export, reports, schedules, health) in original
route order. Rewrote the module docstring to describe the finished
package shape instead of the Task-1-only split.

Retargeted the bare _execute patch in test_field_values_returns_distinct_values
to nx_lib.views.reporting.catalog._execute; verified with a break/revert
probe. Also found and fixed a pre-existing miss from the Task-3 run.py
commit: test_reporting_metrics_routes.py's
test_run_with_metric_returns_aggregated_rows still patched the bare
nx_lib.views.reporting._execute for api_run, which moved to run.py in
an earlier commit of this task -- retargeted to
nx_lib.views.reporting.run._execute and updated its stale comment.

Full non-e2e reporting suite (test_reporting_routes.py,
test_reporting_ai_routes.py, test_reporting_metrics_routes.py,
test_generali_reporting_routes.py, and the reporting unit tests) is
green: 326 passed.
